@@ -2,76 +2,80 @@
   "name": "reception-app",
   "version": "1.0",
   "scripts": {
-    "start": "webpack-dev-server -d -w --inline",
-    "start:remote": "webpack-dev-server -d -w --inline --host 0.0.0.0",
-    "lint": "eslint src/**/*.ts",
-    "build": "webpack -p",
-    "webpack-stat-dev": "webpack -d --config webpack.config.stat.ts",
-    "webpack-stat": "webpack -p --config webpack.config.stat.ts"
+    "start": "webpack serve",
+    "start:remote": "webpack serve --inline --host 0.0.0.0",
+    "lint": "eslint -c eslint.config.mjs src/**/*.{js,jsx,ts,tsx}",
+    "build": "webpack --mode=production"
   },
   "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.14.16",
-    "@mui/material": "^5.14.16",
-    "@mui/styles": "^5.14.16",
-    "classnames": "^2.3.1",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
+    "@mui/icons-material": "^5.16.6",
+    "@mui/material": "^5.16.6",
+    "@mui/styles": "^5.16.6",
+    "classnames": "^2.5.1",
     "jsqr": "^1.4.0",
     "localforage": "^1.10.0",
-    "qrcode": "^1.4.4",
-    "react": "^18.2.0",
-    "react-collapsed": "^4.1.1",
-    "react-dom": "^18.2.0",
+    "process": "^0.11.10",
+    "qrcode": "^1.5.3",
+    "react": "^18.3.1",
+    "react-collapsed": "^4.1.2",
+    "react-dom": "^18.3.1",
     "react-modal": "^3.16.1",
-    "react-redux": "^8.1.3",
-    "react-router-dom": "^6.3.0",
-    "redux": "^4.2.1",
-    "redux-actions": "^2.6.5",
-    "redux-saga": "^1.2.3",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.26.0",
+    "redux": "^5.0.1",
+    "redux-actions": "^3.0.3",
+    "redux-saga": "^1.3.0",
     "typesafe-actions": "4.4.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.12.1",
-    "@babel/plugin-proposal-optional-chaining": "^7.12.7",
-    "@types/copy-webpack-plugin": "^5.0.0",
-    "@types/html-webpack-plugin": "^3.2.1",
-    "@types/node": "^12.7.3",
-    "@types/qrcode": "^1.4.0",
-    "@types/react": "17.0.2",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-modal": "^3.12.0",
-    "@types/react-redux": "^7.1.24",
-    "@types/react-router-dom": "^4.3.5",
-    "@types/terser-webpack-plugin": "^1.2.1",
-    "@types/webpack": "^4.39.1",
-    "@types/webpack-bundle-analyzer": "^2.13.3",
-    "@types/webpack-dev-server": "^3.1.7",
-    "@types/webpack-env": "^1.14.0",
-    "@types/workbox-webpack-plugin": "^5.1.5",
-    "@typescript-eslint/eslint-plugin": "^4.11.1",
-    "@typescript-eslint/parser": "^4.11.1",
-    "babel-loader": "^8.0.6",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.0.4",
-    "css-loader": "^5.0.1",
-    "eslint": "^7.16.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-prettier": "^3.3.0",
-    "eslint-plugin-react": "^7.26.1",
-    "fork-ts-checker-webpack-plugin": "^1.5.0",
-    "html-webpack-plugin": "^3.2.0",
-    "prettier": "^2.2.1",
-    "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "1.4.1",
-    "ts-loader": "^6.0.4",
-    "ts-node": "^8.3.0",
-    "typescript": "^4.1.3",
-    "webpack": "^4.39.3",
-    "webpack-bundle-analyzer": "^3.6.0",
-    "webpack-cli": "^3.3.7",
-    "webpack-dev-server": "^3.8.0",
-    "workbox-sw": "^6.1.5",
-    "workbox-webpack-plugin": "^6.0.2"
+    "@babel/core": "^7.25.2",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
+    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
+    "@types/copy-webpack-plugin": "^10.1.0",
+    "@types/html-webpack-plugin": "^3.2.9",
+    "@types/node": "^22.1.0",
+    "@types/qrcode": "^1.5.5",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-modal": "^3.16.3",
+    "@types/react-redux": "^7.1.33",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/terser-webpack-plugin": "^5.2.0",
+    "@types/webpack": "^5.28.5",
+    "@types/webpack-bundle-analyzer": "^4.7.0",
+    "@types/webpack-env": "^1.18.5",
+    "@types/workbox-webpack-plugin": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
+    "babel-loader": "^9.1.3",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^12.0.2",
+    "css-loader": "^7.1.2",
+    "eslint": "^9.8.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin": "^1.0.1",
+    "eslint-plugin-html": "^8.1.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "fork-ts-checker-webpack-plugin": "^9.0.2",
+    "globals": "^15.9.0",
+    "html-webpack-plugin": "^5.6.0",
+    "prettier": "^3.3.3",
+    "style-loader": "^4.0.0",
+    "terser-webpack-plugin": "^5.3.10",
+    "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.4",
+    "typescript-eslint": "^8.0.0",
+    "webpack": "^5.93.0",
+    "webpack-bundle-analyzer": "^4.10.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4",
+    "workbox-sw": "^7.1.0",
+    "workbox-webpack-plugin": "^7.1.0"
   }
 }
