@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import './index.css';
 import App from './components/pages/App';
@@ -7,11 +7,12 @@ import configureStore from './store';
 import * as serviceWorker from './serviceWorker';
 import { SWUpdateDialog } from './components/organisms/SWUpdateDialog';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container as Element);
+root.render(
   <Provider store={configureStore()}>
     <App />
-  </Provider>,
-  document.getElementById('root'),
+  </Provider>
 );
 
 if ((module as any).hot) {
@@ -24,7 +25,9 @@ serviceWorker.register({
   },
   onUpdate: (registration) => {
     if (registration.waiting) {
-      ReactDOM.render(<SWUpdateDialog registration={registration} />, document.querySelector('.SW-update-dialog'));
+      const container = document.querySelector('.SW-update-dialog');
+      const updater = createRoot(container as Element);
+      updater.render(<SWUpdateDialog registration={registration} />);
     }
   },
 });
