@@ -27,7 +27,7 @@ type ComponentProps = ReturnType<typeof mapStateToProps>;
 type ActionProps = typeof mapDispatchToProps;
 
 type PropsType = ComponentProps & ActionProps;
-const App: React.SFC<PropsType> = (props: PropsType) => {
+const App: React.FC<PropsType> = (props: PropsType) => {
   const classes = useStyles();
   // 表示対象の入場者
   const [dispVisitorList, setDispVisitorList] = React.useState<Visitor[]>([]);
